Export deserializer functions
