+++ src/deserializer/mod.ts
@@ -6,6 +6,8 @@ export * from './read-uc-list.js';
 export * from './sync-ucd-reader.js';
 export * from './ucd-entity-handler.js';
 export * from './ucd-entity-reader.js';
+export * from './ucd-errors.js';
 export * from './ucd-read-non-finite.js';
 export * from './ucd-reader.js';
+export * from './ucd-rx-value.js';
 export * from './ucd-rx.js';
