+++ src/deserializer/ucd-rx.ts
@@ -1,3 +1,5 @@
+import { noop } from '@proc7ts/primitives';
+
 export interface UcdRx {
   _: UcdValueRx;
 
@@ -20,3 +22,23 @@ export interface UcdMapRx {
   for(key: PropertyKey): UcdRx | undefined;
   end(): void;
 }
+
+export const UCD_OPAQUE_RX = {
+  _: {
+    map: {
+      for(_key: string): UcdRx {
+        return UCD_OPAQUE_RX;
+      },
+      end: noop,
+    },
+    nls(): UcdRx {
+      return UCD_OPAQUE_RX;
+    },
+    any(_value: unknown): 1 {
+      return 1;
+    },
+  },
+  lst(): 1 {
+    return 1;
+  },
+} satisfies UcdRx;
