+++ src/syntax/mod.ts
@@ -1,4 +1,5 @@
 export * from './print-uc-token.js';
+export * from './trim-uc-tokens-tail.js';
 export * from './uc-token-kind.js';
 export * from './uc-token.js';
 export * from './uc-tokenizer-stream.js';
